+++ packages/client/src/gen/video/sfu/models/models.ts
@@ -627,6 +627,49 @@ export interface AppleState {
   isLowPowerModeEnabled: boolean;
 }
 /**
+ * PerformanceStats represents the encoding/decoding statistics for a track.
+ *
+ * @generated from protobuf message stream.video.sfu.models.PerformanceStats
+ */
+export interface PerformanceStats {
+  /**
+   * the type of the track (e.g., video, audio, screen share)
+   *
+   * @generated from protobuf field: stream.video.sfu.models.TrackType track_type = 1;
+   */
+  trackType: TrackType;
+  /**
+   * the codec used for the track
+   *
+   * @generated from protobuf field: stream.video.sfu.models.Codec codec = 2;
+   */
+  codec?: Codec;
+  /**
+   * the average encode/decode time in ms
+   *
+   * @generated from protobuf field: float avg_frame_time_ms = 3;
+   */
+  avgFrameTimeMs: number;
+  /**
+   * the average fps for the track
+   *
+   * @generated from protobuf field: float avg_fps = 4;
+   */
+  avgFps: number;
+  /**
+   * the track dimensions
+   *
+   * @generated from protobuf field: stream.video.sfu.models.VideoDimension video_dimension = 5;
+   */
+  videoDimension?: VideoDimension;
+  /**
+   * the target bitrate for the track, only for published tracks
+   *
+   * @generated from protobuf field: int32 target_bitrate = 6;
+   */
+  targetBitrate: number;
+}
+/**
  * @generated from protobuf enum stream.video.sfu.models.PeerType
  */
 export enum PeerType {
@@ -1636,3 +1679,44 @@ class AppleState$Type extends MessageType<AppleState> {
  * @generated MessageType for protobuf message stream.video.sfu.models.AppleState
  */
 export const AppleState = new AppleState$Type();
+// @generated message type with reflection information, may provide speed optimized methods
+class PerformanceStats$Type extends MessageType<PerformanceStats> {
+  constructor() {
+    super('stream.video.sfu.models.PerformanceStats', [
+      {
+        no: 1,
+        name: 'track_type',
+        kind: 'enum',
+        T: () => [
+          'stream.video.sfu.models.TrackType',
+          TrackType,
+          'TRACK_TYPE_',
+        ],
+      },
+      { no: 2, name: 'codec', kind: 'message', T: () => Codec },
+      {
+        no: 3,
+        name: 'avg_frame_time_ms',
+        kind: 'scalar',
+        T: 2 /*ScalarType.FLOAT*/,
+      },
+      { no: 4, name: 'avg_fps', kind: 'scalar', T: 2 /*ScalarType.FLOAT*/ },
+      {
+        no: 5,
+        name: 'video_dimension',
+        kind: 'message',
+        T: () => VideoDimension,
+      },
+      {
+        no: 6,
+        name: 'target_bitrate',
+        kind: 'scalar',
+        T: 5 /*ScalarType.INT32*/,
+      },
+    ]);
+  }
+}
+/**
+ * @generated MessageType for protobuf message stream.video.sfu.models.PerformanceStats
+ */
+export const PerformanceStats = new PerformanceStats$Type();