+++ packages/client/src/stats/rtc/mediaDevices.ts
@@ -24,17 +24,18 @@ if (
     const original = target[method];
     if (!original) continue;
 
+    let mark = 0;
     target[method] = async function tracedMethod(
       constraints: MediaStreamConstraints,
     ) {
-      const tag = `navigator.mediaDevices.${method}`;
+      const tag = `navigator.mediaDevices.${method}.${mark++}`;
       trace(tag, constraints);
       try {
         const stream = await original.call(target, constraints);
-        trace(`${tag}OnSuccess`, dumpStream(stream));
+        trace(`${tag}.OnSuccess`, dumpStream(stream));
         return stream;
       } catch (err) {
-        trace(`${tag}OnFailure`, (err as Error).name);
+        trace(`${tag}.OnFailure`, (err as Error).name);
         throw err;
       }
     };